feat: support Base L1 tx format in batcher recent-tx scanner
Add an --l1-tx-format flag (ethereum|base) to the batcher so an appchain
settling to Base can scan recent L1 txs whose blocks contain deposit and
EIP-8130 entries that standard Ethereum full-block decoding rejects.

Move L1RpcProvider into base-common-network as a format-aware enum wrapping
either an Ethereum or Base RootProvider, and route the recent-tx scanner
through it. Validate that the base format is only used with calldata DA and
no forced blobs, since a Base parent chain has no blob DA endpoint.
