+++ crates/batcher/service/src/lib.rs
@@ -7,6 +7,8 @@
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 
+pub use base_common_genesis::L1TxFormat;
+
 mod config;
 pub use config::BatcherConfig;
