+++ crates/batcher/service/src/recent_txs.rs
@@ -2,10 +2,10 @@
 
 use std::collections::HashMap;
 
+use alloy_consensus::{Transaction, TxEnvelope, transaction::SignerRecoverable};
 use alloy_primitives::Address;
-use alloy_provider::{Provider, RootProvider};
-use alloy_rpc_types_eth::{BlockNumberOrTag, TransactionTrait};
 use base_common_genesis::RollupConfig;
+use base_common_network::{L1RpcProvider, L1RpcProviderError};
 use base_protocol::{Batch, BatchReader, BlockInfo, Channel, ChannelId, Frame};
 use futures::StreamExt;
 use tracing::{debug, info};
@@ -20,6 +20,12 @@ pub const MAX_CHECK_RECENT_TXS_DEPTH: u64 = 128;
 /// Bounds peak memory to roughly this many full L1 blocks while still
 /// achieving significant speedup over sequential fetching.
 pub const SCAN_FETCH_CONCURRENCY: usize = 16;
+
+#[derive(Debug)]
+struct RecentL1Block {
+    info: BlockInfo,
+    transactions: Vec<TxEnvelope>,
+}
 
 /// Scans recent L1 blocks on startup to find the highest submitted L2 block.
 ///
@@ -46,7 +52,7 @@ impl RecentTxScanner {
     /// are never completed and will be silently missed. The caller should treat
     /// the result as a best-effort lower bound, not a guarantee.
     pub async fn highest_submitted_l2_block(
-        l1_provider: &RootProvider,
+        l1_provider: &L1RpcProvider,
         batcher_address: Address,
         batch_inbox: Address,
         depth: u64,
@@ -77,14 +83,7 @@ impl RecentTxScanner {
         let block_stream = futures::stream::iter(scan_start..=current_l1)
             .map(|block_num| {
                 let provider = l1_provider.clone();
-                async move {
-                    let block = provider
-                        .get_block_by_number(BlockNumberOrTag::Number(block_num))
-                        .full()
-                        .await
-                        .map_err(|e| eyre::eyre!("failed to fetch L1 block {block_num}: {e}"))?;
-                    eyre::Ok((block_num, block))
-                }
+                Self::fetch_recent_l1_block(provider, block_num)
             })
             .buffered(SCAN_FETCH_CONCURRENCY);
         futures::pin_mut!(block_stream);
@@ -99,24 +98,17 @@ impl RecentTxScanner {
                 }
             };
 
-            let block_info = BlockInfo {
-                hash: block.header.hash,
-                number: block_num,
-                parent_hash: block.header.inner.parent_hash,
-                timestamp: block.header.inner.timestamp,
-            };
-
-            for tx in block.transactions.txns() {
-                if tx.inner.signer() != batcher_address {
+            for tx in block.transactions {
+                if tx.to() != Some(batch_inbox) {
                     continue;
                 }
-                if tx.inner.to() != Some(batch_inbox) {
+                if tx.recover_signer().ok() != Some(batcher_address) {
                     continue;
                 }
 
                 // Only parse calldata (version-0) frames. Blob transactions have
                 // empty or absent calldata and will fail parse_frames gracefully.
-                let frames = match Frame::parse_frames(tx.inner.input()) {
+                let frames = match Frame::parse_frames(tx.input()) {
                     Ok(f) => f,
                     Err(_) => continue,
                 };
@@ -124,8 +116,8 @@ impl RecentTxScanner {
                 for frame in frames {
                     let channel = channels
                         .entry(frame.id)
-                        .or_insert_with(|| Channel::new(frame.id, block_info));
-                    if let Err(e) = channel.add_frame(frame, block_info) {
+                        .or_insert_with(|| Channel::new(frame.id, block.info));
+                    if let Err(e) = channel.add_frame(frame, block.info) {
                         debug!(error = %e, "ignoring rejected batcher frame during recent tx scan");
                     }
                 }
@@ -136,7 +128,7 @@ impl RecentTxScanner {
                 channels.iter().filter(|(_, ch)| ch.is_ready()).map(|(id, _)| *id).collect();
             for id in complete_ids {
                 if let Some(ch) = channels.remove(&id) {
-                    Self::decode_channel(&ch, block_info.timestamp, rollup_config, &mut highest_l2);
+                    Self::decode_channel(&ch, block.info.timestamp, rollup_config, &mut highest_l2);
                 }
             }
         }
@@ -148,6 +140,27 @@ impl RecentTxScanner {
         }
 
         Ok(highest_l2)
+    }
+
+    async fn fetch_recent_l1_block(
+        provider: L1RpcProvider,
+        block_num: u64,
+    ) -> eyre::Result<(u64, Option<RecentL1Block>)> {
+        let block = match provider.header_and_transactions_by_number(block_num).await {
+            Ok((header, transactions)) => {
+                let info = BlockInfo {
+                    hash: header.hash_slow(),
+                    number: block_num,
+                    parent_hash: header.parent_hash,
+                    timestamp: header.timestamp,
+                };
+                Some(RecentL1Block { info, transactions })
+            }
+            Err(L1RpcProviderError::BlockNotFound(_)) => None,
+            Err(e) => return Err(eyre::eyre!("failed to fetch L1 block {block_num}: {e}")),
+        };
+
+        Ok((block_num, block))
     }
 
     /// Decodes all batches from a complete channel and updates `highest_l2` with
@@ -176,11 +189,24 @@ impl RecentTxScanner {
 
 #[cfg(test)]
 mod tests {
+    use alloy_consensus::{
+        SignableTransaction, TxEip1559, TxEnvelope, transaction::Recovered,
+        transaction::TransactionInfo,
+    };
     use alloy_eips::eip1898::BlockNumHash;
-    use alloy_primitives::B256;
+    use alloy_primitives::{Address, B256, TxKind, U256};
+    use alloy_provider::RootProvider;
     use alloy_rlp::Encodable;
+    use alloy_rpc_types_eth::Transaction as RpcTransaction;
+    use alloy_signer::SignerSync;
+    use alloy_signer_local::PrivateKeySigner;
     use base_common_genesis::{ChainGenesis, RollupConfig};
-    use base_protocol::{Batch, BlockInfo, Channel, ChannelId, Frame, SingleBatch};
+    use base_common_network::{Base, L1RpcProvider};
+    use base_protocol::{
+        Batch, BlockInfo, Channel, ChannelId, DERIVATION_VERSION_0, Frame, SingleBatch,
+    };
+    use httpmock::{HttpMockRequest, HttpMockResponse, Method::POST, MockServer};
+    use serde_json::{Value, json};
 
     use super::RecentTxScanner;
 
@@ -226,6 +252,189 @@ mod tests {
         let frame = Frame { id, number: 0, data: payload, is_last: true };
         channel.add_frame(frame, block_info).expect("frame must be accepted");
         channel
+    }
+
+    fn json_rpc_response(req: &HttpMockRequest, result: Value) -> String {
+        let id = serde_json::from_slice::<Value>(&req.body_vec())
+            .ok()
+            .and_then(|body| body.get("id").cloned())
+            .unwrap_or(Value::Null);
+        json!({ "jsonrpc": "2.0", "id": id, "result": result }).to_string()
+    }
+
+    fn mock_rpc(server: &MockServer, method: &'static str, result: Value) {
+        server.mock(move |when, then| {
+            when.method(POST).path("/").json_body_includes(format!(r#"{{"method":"{method}"}}"#));
+            then.respond_with(move |req| {
+                HttpMockResponse::builder()
+                    .status(200)
+                    .header("content-type", "application/json")
+                    .body(json_rpc_response(req, result.clone()))
+                    .build()
+            });
+        });
+    }
+
+    fn deposit_tx_json() -> Value {
+        json!({
+            "type": "0x7e",
+            "hash": "0x096c03d72acb06339c9c7860d1c36b6451932ec0ff16fd34aa9e30a73a245e13",
+            "nonce": "0x0",
+            "blockHash": "0x1111111111111111111111111111111111111111111111111111111111111111",
+            "blockNumber": "0x2a",
+            "transactionIndex": "0x0",
+            "from": "0xdeaddeaddeaddeaddeaddeaddeaddeaddead0001",
+            "to": "0x4200000000000000000000000000000000000015",
+            "value": "0x0",
+            "gasPrice": "0x0",
+            "gas": "0xf4240",
+            "input": "0x",
+            "v": "0x0",
+            "r": "0x0",
+            "s": "0x0",
+            "sourceHash": "0x990d7122a1f121f3a6bc45723e28f4921c269037a77e77ffee3c8585136d1a92",
+            "mint": "0x0",
+            "depositReceiptVersion": "0x1"
+        })
+    }
+
+    fn eip8130_tx_json() -> Value {
+        json!({
+            "type": "0x7d",
+            "hash": "0x4242424242424242424242424242424242424242424242424242424242424242",
+            "blockHash": "0x1111111111111111111111111111111111111111111111111111111111111111",
+            "blockNumber": "0x2a",
+            "transactionIndex": "0x1",
+            "from": "0x0000000000000000000000000000000000000011",
+            "gasPrice": "0x12a05f200",
+            "tx": {
+                "chainId": 8453,
+                "sender": "0x0000000000000000000000000000000000000011",
+                "nonceKey": "0x0",
+                "nonceSequence": 7,
+                "expiry": 0,
+                "maxPriorityFeePerGas": "0x3b9aca00",
+                "maxFeePerGas": "0x12a05f200",
+                "gasLimit": 1_000_000,
+                "accountChanges": [],
+                "calls": [],
+                "payer": null
+            },
+            "senderAuth": format!("0x{}", "ab".repeat(32)),
+            "payerAuth": "0x"
+        })
+    }
+
+    fn block_with_txs(txs: Vec<Value>) -> Value {
+        json!({
+            "hash": "0x1111111111111111111111111111111111111111111111111111111111111111",
+            "parentHash": "0x2222222222222222222222222222222222222222222222222222222222222222",
+            "sha3Uncles": "0x1dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347",
+            "miner": "0x0000000000000000000000000000000000000000",
+            "stateRoot": "0x3333333333333333333333333333333333333333333333333333333333333333",
+            "transactionsRoot": "0x56e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421",
+            "receiptsRoot": "0x56e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421",
+            "logsBloom": format!("0x{}", "00".repeat(256)),
+            "difficulty": "0x0",
+            "number": "0x2a",
+            "gasLimit": "0x1c9c380",
+            "gasUsed": "0x0",
+            "timestamp": "0x3f2",
+            "extraData": "0x",
+            "mixHash": "0x0000000000000000000000000000000000000000000000000000000000000000",
+            "nonce": "0x0000000000000000",
+            "baseFeePerGas": "0x1",
+            "transactions": txs,
+            "uncles": [],
+            "withdrawals": [],
+            "blobGasUsed": "0x0",
+            "excessBlobGas": "0x0"
+        })
+    }
+
+    fn frame_input_for_l2_timestamp(timestamp: u64) -> Vec<u8> {
+        let batch = SingleBatch { timestamp, ..Default::default() };
+        let frame = Frame::new([9u8; 16], 0, encode_single_batch(&batch), true);
+        let mut input = vec![DERIVATION_VERSION_0];
+        input.extend_from_slice(&frame.encode());
+        input
+    }
+
+    fn signed_batcher_tx_json(
+        signer: &PrivateKeySigner,
+        to: Address,
+        input: Vec<u8>,
+        index: u64,
+    ) -> Value {
+        let tx = TxEip1559 {
+            chain_id: 8453,
+            nonce: index,
+            gas_limit: 1_000_000,
+            max_fee_per_gas: 1_000_000_000,
+            max_priority_fee_per_gas: 1,
+            to: TxKind::Call(to),
+            value: U256::ZERO,
+            input: input.into(),
+            access_list: Default::default(),
+        };
+        let signature = signer.sign_hash_sync(&tx.signature_hash()).expect("tx signs");
+        let envelope: TxEnvelope = tx.into_signed(signature).into();
+        let rpc_tx = RpcTransaction::from_transaction(
+            Recovered::new_unchecked(envelope, signer.address()),
+            TransactionInfo {
+                hash: None,
+                index: Some(index),
+                block_hash: Some(B256::repeat_byte(0x11)),
+                block_number: Some(0x2a),
+                base_fee: Some(1),
+                block_timestamp: Some(0x3f2),
+            },
+        );
+        serde_json::to_value(rpc_tx).expect("transaction serializes")
+    }
+
+    async fn run_recent_scan(provider: L1RpcProvider, batcher: Address, inbox: Address) -> u64 {
+        let cfg = test_rollup_config(1000, 1000, 2);
+        RecentTxScanner::highest_submitted_l2_block(&provider, batcher, inbox, 1, &cfg)
+            .await
+            .expect("scan succeeds")
+            .expect("frame found")
+    }
+
+    #[tokio::test]
+    async fn highest_submitted_l2_block_base_block_drops_non_standard_txs() {
+        let server = MockServer::start_async().await;
+        let signer = PrivateKeySigner::from_slice(&[1u8; 32]).expect("valid signer");
+        let inbox = Address::repeat_byte(0x88);
+        let tx = signed_batcher_tx_json(&signer, inbox, frame_input_for_l2_timestamp(1010), 2);
+        mock_rpc(&server, "eth_blockNumber", json!("0x2a"));
+        mock_rpc(
+            &server,
+            "eth_getBlockByNumber",
+            block_with_txs(vec![deposit_tx_json(), eip8130_tx_json(), tx]),
+        );
+
+        let url = server.url("/").parse().expect("valid url");
+        let provider = L1RpcProvider::Base(RootProvider::<Base>::new_http(url));
+        let highest = run_recent_scan(provider, signer.address(), inbox).await;
+
+        assert_eq!(highest, 1005);
+    }
+
+    #[tokio::test]
+    async fn highest_submitted_l2_block_ethereum_path_recovers_frame() {
+        let server = MockServer::start_async().await;
+        let signer = PrivateKeySigner::from_slice(&[1u8; 32]).expect("valid signer");
+        let inbox = Address::repeat_byte(0x88);
+        let tx = signed_batcher_tx_json(&signer, inbox, frame_input_for_l2_timestamp(1012), 0);
+        mock_rpc(&server, "eth_blockNumber", json!("0x2a"));
+        mock_rpc(&server, "eth_getBlockByNumber", block_with_txs(vec![tx]));
+
+        let url = server.url("/").parse().expect("valid url");
+        let provider = L1RpcProvider::Ethereum(RootProvider::new_http(url));
+        let highest = run_recent_scan(provider, signer.address(), inbox).await;
+
+        assert_eq!(highest, 1006);
     }
 
     // ── decode_channel tests ─────────────────────────────────────────────────
